Completed raster kernel, needs some changes (some indexes are flipped)
