+++ cuda/src/CudaTriangle.cuh
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "Triangle.h"
 #include "MathHelpers.cuh"
+#include "Triangle.h"
 #include <cuda_runtime.h>
 #include <vector>
 
@@ -22,13 +22,15 @@ struct CudaTriangle {
   float3 colB;
   float3 colC;
 
-	float3 projA;
-	float3 projB;
-	float3 projC;
+  float3 projA;
+  float3 projB;
+  float3 projC;
+
+	bool isProjected;
 };
 
 CudaTriangle *triangleToCudaTriangle(const std::vector<Triangle> &triangles);
-void freeCudaTriangles(CudaTriangle* &deviceTriangles);
+void freeCudaTriangles(CudaTriangle *&deviceTriangles);
 __device__ bool isPointInside(CudaTriangle *triangle, float3 p);
 __device__ float3 getBarycentric(CudaTriangle *triangle, float3 p);
 __device__ float3 getBarycentricNormal(CudaTriangle *triangle, float3 b);
